+++ tranformers_test.py
@@ -1,0 +1,4 @@
+from transformers import BlipProcessor
+from sentence_transformers import SentenceTransformer
+
+print("OK")